feat: Convert SHA256 Hash to hexadecimal number and update ApkScanner and AppScanner
