+++ app/src/main/java/tech/projectmatris/antimalwareapp/scanners/ApkScanner.java
@@ -26,10 +26,13 @@ import java.lang.ref.WeakReference;
 import java.nio.MappedByteBuffer;
 import java.nio.channels.FileChannel;
 import java.util.ArrayList;
+import java.util.List;
 
 import androidx.annotation.RequiresApi;
 import tech.projectmatris.antimalwareapp.R;
 import tech.projectmatris.antimalwareapp.activities.AppDetails;
+import tech.projectmatris.antimalwareapp.utils.AppConstants;
+import tech.projectmatris.antimalwareapp.utils.Sha256HashExtractor;
 
 public class ApkScanner extends AsyncTask<Void, String, Void> {
     private WeakReference<Context> contextRef;
@@ -46,6 +49,7 @@ public class ApkScanner extends AsyncTask<Void, String, Void> {
     private float predictionScore;
 
     private ArrayList<String> appPermissionsList= new ArrayList<>();
+    private String sha256Hash;
 
     public ApkScanner(Context context, String path) {
         contextRef = new WeakReference<>(context);
@@ -114,6 +118,11 @@ public class ApkScanner extends AsyncTask<Void, String, Void> {
                 return null;
             }
 
+            List<String> signatures = Sha256HashExtractor.getSignatures(contextRef.get().getPackageManager(), apkMeta.getPackageName());
+            if (signatures != null && !signatures.isEmpty()) {
+                sha256Hash = signatures.get(0);
+            }
+
             float[][] outputVal = new float[1][1];
 
             // Run the model
@@ -141,6 +150,7 @@ public class ApkScanner extends AsyncTask<Void, String, Void> {
         intent.putExtra("prediction",predictionScore);
         intent.putExtra("scan_mode","apk_scan");
         intent.putStringArrayListExtra("permissionList",appPermissionsList);
+        intent.putExtra(AppConstants.SHA_256_HASH, sha256Hash);
         contextRef.get().startActivity(intent);
     }
