+++ app/src/main/java/tech/projectmatris/antimalwareapp/scanners/AppScanner.java
@@ -34,6 +34,8 @@ import android.os.Build;
 
 import tech.projectmatris.antimalwareapp.R;
 import tech.projectmatris.antimalwareapp.activities.AppDetails;
+import tech.projectmatris.antimalwareapp.utils.AppConstants;
+import tech.projectmatris.antimalwareapp.utils.Sha256HashExtractor;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -52,6 +54,7 @@ import java.nio.MappedByteBuffer;
 import java.nio.channels.FileChannel;
 import java.util.ArrayList;
 import java.util.Date;
+import java.util.List;
 
 import androidx.core.app.NotificationCompat;
 import androidx.core.app.NotificationManagerCompat;
@@ -72,6 +75,7 @@ public class AppScanner extends AsyncTask<Void, String, Void> {
     private boolean withSysApps;
 
     private ArrayList<String> appPermissionsList= new ArrayList<>();
+    private String sha256Hash;
 
     public AppScanner(Context context, String pkgName) {
         contextRef = new WeakReference<>(context);
@@ -142,6 +146,11 @@ public class AppScanner extends AsyncTask<Void, String, Void> {
                     return null;
                 }
 
+                List<String> signatures = Sha256HashExtractor.getSignatures(pm, packageName);
+                if (signatures != null && !signatures.isEmpty()) {
+                    sha256Hash = signatures.get(0);
+                }
+
                 float[][] outputVal = new float[1][1];
 
                 // Run the model
@@ -181,6 +190,7 @@ public class AppScanner extends AsyncTask<Void, String, Void> {
                 intent.putExtra("prediction", predictionScore);
                 intent.putExtra("scan_mode", "realtime_scan");
                 intent.putStringArrayListExtra("permissionList", appPermissionsList);
+                intent.putExtra(AppConstants.SHA_256_HASH, sha256Hash);
                 PendingIntent contentIntent = PendingIntent.getActivity(contextRef.get(), 1, intent, PendingIntent.FLAG_CANCEL_CURRENT);
                 if (prediction.equalsIgnoreCase(contextRef.get().getString(R.string.safe))) {
                     notificationBuilder.setSmallIcon(R.drawable.ic_notification)
