+++ app/src/main/java/tech/projectmatris/antimalwareapp/scanners/ScannerTask.java
@@ -27,9 +27,7 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.pm.ActivityInfo;
 import android.content.pm.ApplicationInfo;
-import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
-import android.content.pm.Signature;
 import android.content.res.AssetFileDescriptor;
 import android.content.res.AssetManager;
 import android.content.res.Configuration;
@@ -37,18 +35,17 @@ import android.content.res.XmlResourceParser;
 import android.os.AsyncTask;
 import android.os.Build;
 import android.preference.PreferenceManager;
-import android.util.Base64;
 import android.widget.ProgressBar;
 import android.widget.TextView;
 
 import tech.projectmatris.antimalwareapp.R;
 
-import androidx.annotation.NonNull;
 import androidx.core.app.NotificationCompat;
 import androidx.core.app.NotificationManagerCompat;
 
 import tech.projectmatris.antimalwareapp.activities.ResultActivity;
 import tech.projectmatris.antimalwareapp.data.AppInfo;
+import tech.projectmatris.antimalwareapp.utils.Sha256HashExtractor;
 
 
 import org.json.JSONArray;
@@ -232,7 +229,7 @@ public class ScannerTask extends AsyncTask<Void, String, Void> {
                     float inferredValue = outputVal[0][0];
                     app.predictionScore=inferredValue;
                     app.permissionList=appPermissionsList;
-                    List<String> signatures = getSignatures(pm, packageInfo.packageName);
+                    List<String> signatures = Sha256HashExtractor.getSignatures(pm, packageInfo.packageName);
                     if (signatures != null && !signatures.isEmpty()) {
                         app.sha256Hash = signatures.get(0);
                     }
@@ -448,56 +445,4 @@ public class ScannerTask extends AsyncTask<Void, String, Void> {
         editor.putString("lastScan", curDateTime);
         editor.apply();
     }
-
-    /**
-     * Get the list of Signatures used by the application
-     */
-    private static List<String> getSignatures(@NonNull PackageManager pm, @NonNull String packageName) {
-        try {
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
-                PackageInfo packageInfo = pm.getPackageInfo(packageName, PackageManager.GET_SIGNING_CERTIFICATES);
-                if (packageInfo == null || packageInfo.signingInfo == null) {
-                    return null;
-                }
-                if (packageInfo.signingInfo.hasMultipleSigners()) {
-                    return convertSignaturesToStrings(packageInfo.signingInfo.getApkContentsSigners());
-                } else {
-                    return convertSignaturesToStrings(packageInfo.signingInfo.getSigningCertificateHistory());
-                }
-            } else {
-                @SuppressLint("PackageManagerGetSignatures")
-                PackageInfo packageInfo = pm.getPackageInfo(packageName, PackageManager.GET_SIGNATURES);
-                if (packageInfo == null || packageInfo.signatures == null || packageInfo.signatures.length == 0 || packageInfo.signatures[0] == null) {
-                    return null;
-                }
-                return convertSignaturesToStrings(packageInfo.signatures);
-            }
-        } catch (PackageManager.NameNotFoundException e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-
-    private static List<String> convertSignaturesToStrings(Signature[] signatures) {
-        List<String> signaturesList = new ArrayList<>();
-        for (Signature signature: signatures) {
-            if (signature != null) {
-                signaturesList.add(getDigestedSignature(signature));
-            }
-        }
-        return signaturesList;
-    }
-
-    private static String getDigestedSignature(Signature signature) {
-        byte[] signatureArray = signature.toByteArray();
-        try {
-            MessageDigest md = MessageDigest.getInstance("SHA-256");
-            byte[] digest = md.digest(signatureArray);
-            return Base64.encodeToString(digest, Base64.DEFAULT);
-        } catch (NoSuchAlgorithmException e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-
 }
